refactor(website): order version diff fields by config order
Reuse the sequence details page's header grouping/ordering in the version
diff table so fields and header groups appear in the central, config-defined
order (orderOnDetailsPage) instead of arbitrary insertion order.

- Extract groupTableDataByHeader() shared helper from getDataTableData
- Carry orderOnDetailsPage through FieldComparison
- Use the shared helper in DiffTable
